moved chart-init into common-scripts reorganization of project added javascript gulp tasks
