+++ package.json
@@ -9,6 +9,9 @@
     "gulp-clip-empty-files": "^0.1.1",
     "gulp-minify-css": "^1.2.1",
     "gulp-rename": "^1.2.2",
-    "gulp-sass": "^2.0.1"
+    "gulp-sass": "^2.0.1",
+    "gulp-concat": "~2.6.0",
+    "gulp-sourcemaps": "~1.6.0",
+    "gulp-uglify": "~1.5.3"
   }
 }
