--- web/src/stylesheets/FlatLab/404.scss
@@ -1,45 +1,0 @@
-
-/*404 page*/
-
-.body-404 {
-	background: #18d4cb;
-	color: #fff;
-}
-
-.error-wrapper {
-	text-align: center;
-	margin-top: 10%;
-}
-
-.error-wrapper .icon-404 {
-	background: url("../img/404_icon.png") no-repeat;
-	width: 289px;
-	height: 274px;
-	display: inline-block;
-	margin-left: 30px;
-}
-
-.error-wrapper h1 {
-	font-size: 90px;
-	font-weight: 300;
-	margin: -50px 0 0 0;
-}
-
-.error-wrapper h2 {
-	font-size: 20px;
-	font-weight: 300;
-	margin: 0 0 30px 0;
-}
-
-.error-wrapper p, .error-wrapper p a {
-	font-size: 18px;
-	font-weight: 300;
-}
-
-.error-wrapper p.page-404 {
-	color: #7dfff7;
-}
-
-.error-wrapper p.page-404 a, .error-wrapper p.page-500 a, .error-wrapper p.page-404 a:hover, .error-wrapper p.page-500 a:hover {
-	color: #fff;
-}
