--- web/src/stylesheets/FlatLab/505.scss
@@ -1,19 +1,0 @@
-
-/*500 page*/
-
-.body-500 {
-	background: #8075c6;
-	color: #fff;
-}
-
-.error-wrapper p.page-500 {
-	color: #afa5f1;
-}
-
-.error-wrapper .icon-500 {
-	background: url("../img/500_icon.png") no-repeat;
-	width: 289px;
-	height: 274px;
-	display: inline-block;
-	margin-left: 55px;
-}
