--- web/src/stylesheets/FlatLab/carousel.scss
@@ -1,64 +1,0 @@
-
-/*carousel*/
-
-.carousel-indicators li {
-	background: rgba(0, 0, 0, 0.2);
-	border: none;
-	transition: background-color 0.25s ease 0s;
-	-moz-transition: background-color 0.25s ease 0s;
-	-webkit-transition: background-color 0.25s ease 0s;
-}
-
-.carousel-indicators .active {
-	background: #ff6c60;
-	height: 10px;
-	margin: 1px;
-	width: 10px;
-}
-
-.carousel-indicators.out {
-	bottom: -5px;
-}
-
-.carousel-indicators.out {
-	bottom: -5px;
-}
-
-.carousel-control {
-	color: #999999;
-	text-shadow: none;
-	width: 45px;
-}
-
-.carousel-control i {
-	display: inline-block;
-	height: 25px;
-	left: 50%;
-	margin-left: -10px;
-	margin-top: -10px;
-	position: absolute;
-	top: 50%;
-	width: 20px;
-	z-index: 5;
-}
-
-.carousel-control.left, .carousel-control.right {
-	background: none;
-	filter: none;
-}
-
-.carousel-control:hover, .carousel-control:focus {
-	color: #CCCCCC;
-	opacity: 0.9;
-	text-decoration: none;
-}
-
-.carousel-inner h3 {
-	font-weight: 300;
-	font-size: 16px;
-	margin: 0;
-}
-
-.carousel-inner {
-	margin-bottom: 15px;
-}
