--- web/src/stylesheets/FlatLab/form.scss
@@ -1,93 +1,0 @@
-
-/*form*/
-
-.sm-input {
-	width: 175px;
-}
-
-.form-horizontal.tasi-form .form-group {
-	border-bottom: 1px solid #eff2f7;
-	padding-bottom: 15px;
-	margin-bottom: 15px;
-}
-
-.form-horizontal.tasi-form .form-group:last-child {
-	border-bottom: none;
-	padding-bottom: 0;
-	margin-bottom: 0;
-}
-
-.form-horizontal.tasi-form .form-group .help-block {
-	margin-bottom: 0;
-}
-
-.round-input {
-	border-radius: 500px;
-	-webkit-border-radius: 500px;
-}
-
-.m-bot15 {
-	margin-bottom: 15px;
-}
-
-.form-horizontal.tasi-form .checkbox-inline > input {
-	margin-top: 1px;
-	border: none;
-}
-
-.iconic-input {
-	position: relative;
-}
-
-.iconic-input i {
-	color: #CCCCCC;
-	display: block;
-	font-size: 16px;
-	height: 16px;
-	margin: 8px 5px 8px 10px;
-	position: absolute;
-	text-align: center;
-	width: 16px;
-}
-
-.iconic-input input {
-	padding-left: 30px !important;
-}
-
-.iconic-input.right input {
-	padding-left: 10px !important;
-	padding-right: 30px !important;
-}
-
-.iconic-input.right i {
-	float: right;
-	right: 5px;
-}
-
-input.spinner[type="text"], input.spinner[type="password"], input.spinner[type="datetime"], input.spinner[type="datetime-local"], input.spinner[type="date"], input.spinner[type="month"], input.spinner[type="time"], input.spinner[type="week"], input.spinner[type="number"], input.spinner[type="email"], input.spinner[type="url"], input.spinner[type="search"], input.spinner[type="tel"], input.spinner[type="color"] {
-	background: url("../img/input-spinner.gif") right no-repeat !important;
-}
-
-/*form validation*/
-
-.cmxform .form-group label.error {
-	display: inline;
-	margin: 5px 0;
-	color: #B94A48;
-	font-weight: 400;
-}
-
-input:focus:invalid:focus, textarea:focus:invalid:focus, select:focus:invalid:focus, .cmxform .form-group input.error, .cmxform .form-group textarea.error {
-	border-color: #B94A48 !important;
-}
-
-#signupForm label.error {
-	display: inline;
-	margin: 5px 0;
-	width: auto;
-	color: #B94A48;
-}
-
-.checkbox, .checkbox:hover, .checkbox:focus {
-	border: none;
-}
