--- web/src/stylesheets/FlatLab/invoice.scss
@@ -1,34 +1,0 @@
-
-/*invoice*/
-
-.invoice-list {
-	margin-bottom: 30px;
-}
-
-.invoice-list h4 {
-	font-weight: 300;
-	font-size: 16px;
-}
-
-.invoice-block {
-	text-align: right;
-}
-
-ul.amounts li {
-	background: #f5f5f5;
-	margin-bottom: 5px;
-	padding: 10px;
-	border-radius: 4px;
-	-webkit-border-radius: 4px;
-	font-weight: 300;
-}
-
-.invoice-btn a {
-	font-weight: 300;
-	margin: 0 5px;
-	font-size: 16px;
-}
-
-.corporate-id {
-	margin-bottom: 30px;
-}
