+++ web/src/stylesheets/templates/font-awesome.scss
