+++ web/src/stylesheets/templates/style-responsive.scss
